+++ lib/ui/wellness/WellnessMentalHealthContentWidget.dart
@@ -14,18 +14,21 @@
  * limitations under the License.
  */
 
-import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:illinois/model/Video.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Config.dart';
 import 'package:illinois/service/DeepLink.dart';
 import 'package:illinois/service/Guide.dart';
+import 'package:illinois/ui/settings/SettingsVideoTutorialPanel.dart';
 import 'package:illinois/ui/wellness/WellnessResourcesContentWidget.dart';
+import 'package:illinois/ui/widgets/VideoPlayButton.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
+import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:webview_flutter/webview_flutter.dart' as flutter_webview;
 
 class WellnessMentalHealthContentWidget extends StatefulWidget {
   WellnessMentalHealthContentWidget();
@@ -39,8 +42,7 @@ class _WellnessMentalHealthContentWidgetState extends State<WellnessMentalHealth
   Map<String, List<Map<String, dynamic>>> _sectionLists = <String, List<Map<String, dynamic>>>{};
   List<String> _sections = <String>[];
 
-  bool _isVideoLoading = false;
-  String? _videoUrl;
+  Video? _video;
 
   @override
   void initState() {
@@ -48,7 +50,7 @@ class _WellnessMentalHealthContentWidgetState extends State<WellnessMentalHealth
       Auth2UserPrefs.notifyFavoritesChanged,
       Guide.notifyChanged,
     ]);
-    _initVideo();
+    _loadVideo();
     _buildContentData();
     super.initState();
   }
@@ -85,7 +87,7 @@ class _WellnessMentalHealthContentWidgetState extends State<WellnessMentalHealth
   Widget _buildContentUi() {
     List<Widget> widgetList = <Widget>[];
 
-    if (StringUtils.isNotEmpty(_videoUrl)) {
+    if (_video != null) {
       widgetList.add(_buildContentVideo());
     }
 
@@ -187,30 +189,53 @@ class _WellnessMentalHealthContentWidgetState extends State<WellnessMentalHealth
   }
 
   Widget _buildContentVideo() {
-    if (StringUtils.isEmpty(_videoUrl)) {
+    if (_video == null) {
       return Container();
     }
-    double width = MediaQuery.of(context).size.width;
-    double height = (width * 3) / 4;
-    return Stack(alignment: Alignment.center, children: [
-      Container(
-          height: height,
-          child: flutter_webview.WebView(
-              initialUrl: _videoUrl,
-              javascriptMode: flutter_webview.JavascriptMode.unrestricted,
-              onPageFinished: (url) {
-                setState(() {
-                  _isVideoLoading = false;
-                });
-              })),
-      Visibility(visible: _isVideoLoading, child: CircularProgressIndicator())
-    ]);
+    String? imageUrl = _video!.thumbUrl;
+    bool hasImage = StringUtils.isNotEmpty(imageUrl);
+    final Widget emptyImagePlaceholder = Container(height: 102);
+    return Container(
+        decoration: BoxDecoration(
+            color: Styles().colors?.white,
+            boxShadow: [BoxShadow(color: Styles().colors!.blackTransparent018!, spreadRadius: 1.0, blurRadius: 3.0, offset: Offset(1, 1))],
+            borderRadius: BorderRadius.vertical(bottom: Radius.circular(4))),
+        child: Stack(children: [
+          GestureDetector(
+              onTap: _onTapVideo,
+              child: Semantics(
+                  button: true,
+                  child: Padding(
+                      padding: EdgeInsets.all(16),
+                      child: Stack(alignment: Alignment.center, children: [
+                        hasImage
+                            ? ClipRRect(
+                                borderRadius: BorderRadius.circular(4),
+                                child: Image.network(imageUrl!,
+                                    loadingBuilder: (BuildContext context, Widget child, ImageChunkEvent? loadingProgress) {
+                                  return (loadingProgress == null) ? child : emptyImagePlaceholder;
+                                }))
+                            : emptyImagePlaceholder,
+                        VideoPlayButton()
+                      ])))),
+          Container(color: Styles().colors?.accentColor3, height: 4)
+        ]));
   }
 
-  void _initVideo() {
-    _videoUrl = Config().wellnessMentalHealthVideoUrl;
-    if (StringUtils.isNotEmpty(_videoUrl)) {
-      _isVideoLoading = true;
+  void _loadVideo() {
+    String? videoUrl = Config().wellnessMentalHealthVideoUrl;
+    if (StringUtils.isNotEmpty(videoUrl)) {
+      String title = Localization().getStringEx('panel.wellness.sections.mental_health.video.title', 'Mental Health');
+      _video = Video(
+          videoUrl: videoUrl, ccUrl: Config().wellnessMentalHealthCcUrl, thumbUrl: Config().wellnessMentalHealthThumbUrl, title: title);
+    }
+  }
+
+  void _onTapVideo() {
+    if (_video != null) {
+      Analytics().logSelect(
+          target: 'Mental Health Video', source: widget.runtimeType.toString(), attributes: _video!.analyticsAttributes);
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => SettingsVideoTutorialPanel(videoTutorial: _video!)));
     }
   }
 }