+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Display video thumb in "Mental Health Resources". Play the video in another panel with subtitles [#3164](https://github.com/rokwire/illinois-app/issues/3164).
 - Hook up appointments APIs [#3167](https://github.com/rokwire/illinois-app/issues/3167).
+- Launch privacy policy content when offline [#3052](https://github.com/rokwire/illinois-app/issues/3052).
 ### Fixed
 - Fix MyCalendarSettings toggle buttons [#3083](https://github.com/rokwire/illinois-app/issues/3083).
 - Improve Accessibility for HomeToutWidget [#3026](https://github.com/rokwire/illinois-app/issues/3026).